+++ skforecast/ForecasterAutoreg/tests/test_fit.py
@@ -4,8 +4,9 @@ import pytest
 from pytest import approx
 import numpy as np
 import pandas as pd
-from skforecast.ForecasterAutoreg import ForecasterAutoreg
 from sklearn.linear_model import LinearRegression
+from skforecast.preprocessing import RollingFeatures
+from skforecast.ForecasterAutoreg import ForecasterAutoreg
 
 # Fixtures
 from .fixtures_ForecasterAutoreg import y
@@ -29,7 +30,11 @@ def test_forecaster_y_exog_features_stored():
     """
     Test forecaster stores y and exog features after fitting.
     """
-    forecaster = ForecasterAutoreg(LinearRegression(), lags=3)
+    rolling = RollingFeatures(
+        stats=['ratio_min_max', 'median'], window_sizes=4
+    )
+    forecaster = ForecasterAutoreg(LinearRegression(), lags=3,
+                                   window_features=rolling)
     forecaster.fit(y=y, exog=exog)
 
     exog_in_ = True
@@ -37,13 +42,16 @@ def test_forecaster_y_exog_features_stored():
     exog_names_in_ = ['exog']
     exog_dtypes_in_ = {'exog': exog.dtype}
     X_train_exog_names_out_ = ['exog']
-    X_train_features_names_out_ = ['lag_1', 'lag_2', 'lag_3', 'exog']
+    X_train_window_features_names_out_ = ['roll_ratio_min_max_4', 'roll_median_4']
+    X_train_features_names_out_ = ['lag_1', 'lag_2', 'lag_3', 
+                                   'roll_ratio_min_max_4', 'roll_median_4', 'exog']
     
     assert forecaster.exog_in_ == exog_in_
     assert forecaster.exog_type_in_ == exog_type_in_
     assert forecaster.exog_names_in_ == exog_names_in_
     assert forecaster.exog_dtypes_in_ == exog_dtypes_in_
     assert forecaster.X_train_exog_names_out_ == X_train_exog_names_out_
+    assert forecaster.X_train_window_features_names_out_ == X_train_window_features_names_out_
     assert forecaster.X_train_features_names_out_ == X_train_features_names_out_