+++ skforecast/ForecasterAutoreg/tests/test_set_lags.py
@@ -19,7 +19,6 @@ def test_set_lags_when_lags_argument_is_int():
     assert (forecaster.lags == np.array([1, 2, 3, 4, 5])).all()
     assert forecaster.max_lag == 5
     assert forecaster.window_size == 5
-    assert forecaster.window_size_diff == 5
 
 
 def test_set_lags_when_lags_argument_is_list():
@@ -33,7 +32,6 @@ def test_set_lags_when_lags_argument_is_list():
     assert (forecaster.lags == np.array([1, 2, 3])).all()
     assert forecaster.max_lag == 3
     assert forecaster.window_size == 3
-    assert forecaster.window_size_diff == 3
 
 
 def test_set_lags_when_lags_argument_is_1d_numpy_array():
@@ -47,7 +45,6 @@ def test_set_lags_when_lags_argument_is_1d_numpy_array():
     assert (forecaster.lags == np.array([1, 2, 3])).all()
     assert forecaster.max_lag == 3
     assert forecaster.window_size == 3
-    assert forecaster.window_size_diff == 3
 
 
 def test_set_lags_when_differentiation_is_not_None():
@@ -64,5 +61,4 @@ def test_set_lags_when_differentiation_is_not_None():
 
     assert (forecaster.lags == np.array([1, 2, 3, 4, 5])).all()
     assert forecaster.max_lag == 5
-    assert forecaster.window_size == 5
-    assert forecaster.window_size_diff == 5 + 1
+    assert forecaster.window_size == 5 + 1